backend/analytics-statistic-service/requirements.txt
# ===== Core FastAPI =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1

# ===== Database (SYNC – đúng với psycopg2) =====
psycopg2-binary==2.9.9

# ===== Data processing =====
pandas==2.2.0
numpy==1.26.3

# ===== Visualization / Report =====
matplotlib==3.8.2
reportlab==4.0.9

# ===== HTTP client =====
httpx==0.26.0

# ===== Validation =====
pydantic==2.5.3
pydantic-settings==2.1.0

# ===== Utilities =====
tenacity==8.2.3
